+++ python_modules/dagster/dagster/_core/definitions/run_request.py
@@ -1,7 +1,6 @@
-from collections import defaultdict
 from collections.abc import Mapping, Sequence
 from enum import Enum
-from typing import TYPE_CHECKING, AbstractSet, Any, NamedTuple, Optional, Union  # noqa: UP035
+from typing import TYPE_CHECKING, Any, NamedTuple, Optional, Union
 
 from dagster_shared.serdes import whitelist_for_serdes
 
@@ -21,22 +20,20 @@ from dagster._core.definitions.dynamic_partitions_request import (
 from dagster._core.definitions.events import AssetKey, AssetMaterialization, AssetObservation
 from dagster._core.definitions.partitions.context import partition_loading_context
 from dagster._core.definitions.partitions.partition_key_range import PartitionKeyRange
-from dagster._core.instance import DynamicPartitionsStore
 from dagster._core.storage.dagster_run import DagsterRun, DagsterRunStatus
 from dagster._core.storage.tags import (
     ASSET_PARTITION_RANGE_END_TAG,
     ASSET_PARTITION_RANGE_START_TAG,
     PARTITION_NAME_TAG,
 )
-from dagster._core.types.pagination import PaginatedResults
-from dagster._record import IHaveNew, LegacyNamedTupleMixin, record, record_custom
-from dagster._utils.cached_method import cached_method
+from dagster._record import IHaveNew, LegacyNamedTupleMixin, record_custom
 from dagster._utils.error import SerializableErrorInfo
 from dagster._utils.tags import normalize_tags
 
 if TYPE_CHECKING:
     from dagster._core.definitions.job_definition import JobDefinition
     from dagster._core.definitions.run_config import RunConfig
+    from dagster._core.instance.types import DynamicPartitionsStore
 
 
 @whitelist_for_serdes(old_storage_names={"JobType"})
@@ -181,8 +178,10 @@ class RunRequest(IHaveNew, LegacyNamedTupleMixin):
         dynamic_partitions_requests: Sequence[
             Union[AddDynamicPartitionsRequest, DeleteDynamicPartitionsRequest]
         ],
-        dynamic_partitions_store: Optional[DynamicPartitionsStore],
+        dynamic_partitions_store: Optional["DynamicPartitionsStore"],
     ) -> "RunRequest":
+        from dagster._core.instance.types import DynamicPartitionsStoreAfterRequests
+
         if self.partition_key is None:
             check.failed(
                 "Cannot resolve partition for run request without partition key",
@@ -247,75 +246,6 @@ class RunRequest(IHaveNew, LegacyNamedTupleMixin):
         execute it as a single run instead.
         """
         return self.asset_graph_subset is not None
-
-
-@record
-class DynamicPartitionsStoreAfterRequests(DynamicPartitionsStore):
-    """Represents the dynamic partitions that will be in the contained DynamicPartitionsStore
-    after the contained requests are satisfied.
-    """
-
-    wrapped_dynamic_partitions_store: DynamicPartitionsStore
-    added_partition_keys_by_partitions_def_name: Mapping[str, AbstractSet[str]]
-    deleted_partition_keys_by_partitions_def_name: Mapping[str, AbstractSet[str]]
-
-    @staticmethod
-    def from_requests(
-        wrapped_dynamic_partitions_store: DynamicPartitionsStore,
-        dynamic_partitions_requests: Sequence[
-            Union[AddDynamicPartitionsRequest, DeleteDynamicPartitionsRequest]
-        ],
-    ) -> "DynamicPartitionsStoreAfterRequests":
-        added_partition_keys_by_partitions_def_name: dict[str, set[str]] = defaultdict(set)
-        deleted_partition_keys_by_partitions_def_name: dict[str, set[str]] = defaultdict(set)
-
-        for req in dynamic_partitions_requests:
-            name = req.partitions_def_name
-            if isinstance(req, AddDynamicPartitionsRequest):
-                added_partition_keys_by_partitions_def_name[name].update(set(req.partition_keys))
-            elif isinstance(req, DeleteDynamicPartitionsRequest):
-                deleted_partition_keys_by_partitions_def_name[name].update(set(req.partition_keys))
-            else:
-                check.failed(f"Unexpected request type: {req}")
-
-        return DynamicPartitionsStoreAfterRequests(
-            wrapped_dynamic_partitions_store=wrapped_dynamic_partitions_store,
-            added_partition_keys_by_partitions_def_name=added_partition_keys_by_partitions_def_name,
-            deleted_partition_keys_by_partitions_def_name=deleted_partition_keys_by_partitions_def_name,
-        )
-
-    @cached_method
-    def get_dynamic_partitions(self, partitions_def_name: str) -> Sequence[str]:
-        partition_keys = set(
-            self.wrapped_dynamic_partitions_store.get_dynamic_partitions(partitions_def_name)
-        )
-        added_partition_keys = self.added_partition_keys_by_partitions_def_name.get(
-            partitions_def_name, set()
-        )
-        deleted_partition_keys = self.deleted_partition_keys_by_partitions_def_name.get(
-            partitions_def_name, set()
-        )
-        return list((partition_keys | added_partition_keys) - deleted_partition_keys)
-
-    @cached_method
-    def get_paginated_dynamic_partitions(
-        self, partitions_def_name: str, limit: int, ascending: bool, cursor: Optional[str] = None
-    ) -> PaginatedResults[str]:
-        partition_keys = self.get_dynamic_partitions(partitions_def_name)
-        return PaginatedResults.create_from_sequence(
-            seq=partition_keys, limit=limit, ascending=ascending, cursor=cursor
-        )
-
-    def has_dynamic_partition(self, partitions_def_name: str, partition_key: str) -> bool:
-        return partition_key not in self.deleted_partition_keys_by_partitions_def_name.get(
-            partitions_def_name, set()
-        ) and (
-            partition_key
-            in self.added_partition_keys_by_partitions_def_name.get(partitions_def_name, set())
-            or self.wrapped_dynamic_partitions_store.has_dynamic_partition(
-                partitions_def_name, partition_key
-            )
-        )
 
 
 @whitelist_for_serdes(
